refactor: domain 내의 패키지 재정돈
