+++ src/main/kotlin/racingcar/domain/engine/Engine.kt
@@ -1,4 +1,4 @@
-package racingcar.domain.car.engine
+package racingcar.domain.engine
 
 interface Engine {
     fun run(): Torque
