+++ src/main/kotlin/racingcar/domain/engine/ForceGenerator.kt
@@ -1,4 +1,4 @@
-package racingcar.domain.car.engine
+package racingcar.domain.engine
 
 interface ForceGenerator {
     fun generate(): Force
