+++ src/main/kotlin/racingcar/domain/racing/RacingTrial.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.domain.racing
 
 import racingcar.domain.car.Car
