+++ src/main/kotlin/racingcar/domain/racing/StandardRacingCarGenerator.kt
@@ -1,8 +1,8 @@
-package racingcar.domain
+package racingcar.domain.racing
 
 import racingcar.domain.car.Car
 import racingcar.domain.car.CarName
-import racingcar.domain.car.engine.RacingEngine
+import racingcar.domain.engine.RacingEngine
 
 class StandardRacingCarGenerator : RacingCarGenerator {
     override fun generate(name: CarName): Car {
