+++ src/test/kotlin/racingcar/domain/car/CarLocationTest.kt
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.ValueSource
-import racingcar.domain.car.engine.Torque
+import racingcar.domain.engine.Torque
 
 internal class CarLocationTest {
     @Test
