+++ src/test/kotlin/racingcar/domain/racing/RacingTest.kt
@@ -1,13 +1,12 @@
-package racingcar.domain
+package racingcar.domain.racing
 
-import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
+import org.assertj.core.api.Assertions
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 import racingcar.domain.car.Car
 import racingcar.domain.car.CarName
-import racingcar.domain.car.engine.Engine
-import racingcar.domain.car.engine.Torque
+import racingcar.domain.engine.Engine
+import racingcar.domain.engine.Torque
 
 internal class RacingTest {
     private val dummyEngine = object : Engine {
@@ -17,7 +16,7 @@ internal class RacingTest {
     @ParameterizedTest
     @ValueSource(ints = [0, 1])
     fun `참가 차량이 2대보다 적다면 IllegalArgumentException을 던진다`(numberOfCars: Int) {
-        assertThatIllegalArgumentException()
+        Assertions.assertThatIllegalArgumentException()
             .isThrownBy { Racing(1, generateCars(numberOfCars)) }
             .withMessage("At least two cars must be participated. count='$numberOfCars'")
     }
@@ -25,7 +24,7 @@ internal class RacingTest {
     @ParameterizedTest
     @ValueSource(ints = [0, -1, -38])
     fun `시도 횟수는 1보다 작거나 같으면 IllegalArgumentException을 던진다`(trials: Int) {
-        assertThatIllegalArgumentException()
+        Assertions.assertThatIllegalArgumentException()
             .isThrownBy { Racing(trials, generateCars(2)) }
             .withMessage("The value of trials must be positive value.")
     }
@@ -34,7 +33,7 @@ internal class RacingTest {
     @ValueSource(ints = [1, 10, 51])
     fun `시도 횟수만큼 경주 결과(RacingResults) 내에 경주 시도(RacingTrial)가 생긴다`(trials: Int) {
         val results = Racing(trials, generateCars(2)).start()
-        assertThat(results.trials).hasSize(trials)
+        Assertions.assertThat(results.trials).hasSize(trials)
     }
 
     @ParameterizedTest
@@ -43,7 +42,7 @@ internal class RacingTest {
         val cars = generateCars(numberOfCars)
 
         val results = Racing(1, cars).start()
-        assertThat(results.trials).allMatch { it.records.size == numberOfCars }
+        Assertions.assertThat(results.trials).allMatch { it.records.size == numberOfCars }
     }
 
     private fun generateCars(numberOfCars: Int) = (1..numberOfCars).map { Car(CarName("슈마허"), dummyEngine) }
