+++ src/test/kotlin/racingcar/domain/racing/RecordTest.kt
@@ -1,6 +1,6 @@
-package racingcar.domain
+package racingcar.domain.racing
 
-import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.assertAll
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
@@ -18,8 +18,8 @@ internal class RecordTest {
     fun `기록은 생성될 때의 차의 이름과 위치를 그대로 반환한다`(name: String, location: Int) {
         val record = Record(CarName(name), CarLocation(location))
         assertAll(
-            { assertThat(record.name).isEqualTo(CarName(name)) },
-            { assertThat(record.location).isEqualTo(CarLocation(location)) }
+            { Assertions.assertThat(record.name).isEqualTo(CarName(name)) },
+            { Assertions.assertThat(record.location).isEqualTo(CarLocation(location)) }
         )
     }
 
@@ -28,7 +28,7 @@ internal class RecordTest {
     fun `다른 기록의 위치와 같으면 true를 반환한다`(location: Int) {
         val one = Record("one", location)
         val other = Record("other", location)
-        assertThat(one.isSameLocationAs(other)).isTrue()
+        Assertions.assertThat(one.isSameLocationAs(other)).isTrue()
     }
 
     @ParameterizedTest
@@ -38,7 +38,7 @@ internal class RecordTest {
         "제네시스, 2"
     )
     fun `Record는 객체가 달라도 값이 같으면 동일하다`(name: String, location: Int) {
-        assertThat(Record(name, location)).isEqualTo(Record(CarName(name), CarLocation(location)))
+        Assertions.assertThat(Record(name, location)).isEqualTo(Record(CarName(name), CarLocation(location)))
     }
 
     @ParameterizedTest
@@ -46,6 +46,6 @@ internal class RecordTest {
     fun `다른 기록의 위치와 다르면 false를 반환한다`(oneLocation: Int, otherLocation: Int) {
         val one = Record("one", oneLocation)
         val other = Record("other", otherLocation)
-        assertThat(one.isSameLocationAs(other)).isFalse()
+        Assertions.assertThat(one.isSameLocationAs(other)).isFalse()
     }
 }
