+++ src/test/kotlin/racingcar/domain/racing/StandardRacingCarGeneratorTest.kt
@@ -1,6 +1,6 @@
-package racingcar.domain
+package racingcar.domain.racing
 
-import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
@@ -17,6 +17,6 @@ internal class StandardRacingCarGeneratorTest {
     @ParameterizedTest
     @ValueSource(strings = ["남동민", "ABCDE", "한글다섯자", "12 45", "", "1"])
     fun `차 생산 시, 넣어준 이름과 똑같은 이름의 차가 생성된다`(name: String) {
-        assertThat(generator.generate(CarName(name)).name).isEqualTo(CarName(name))
+        Assertions.assertThat(generator.generate(CarName(name)).name).isEqualTo(CarName(name))
     }
 }
