+++ src/test/kotlin/racingcar/domain/racing/TrialOrderTest.kt
@@ -1,7 +1,6 @@
-package racingcar.domain
+package racingcar.domain.racing
 
-import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
+import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
@@ -10,19 +9,19 @@ internal class TrialOrderTest {
     @ParameterizedTest
     @ValueSource(ints = [1, 48, 203802, Int.MAX_VALUE])
     fun `시도 순서(TrialOrder)의 값을 조회하면 생성 시 입력한 값을 그대로 가지고 있다`(value: Int) {
-        assertThat(TrialOrder(value).value).isEqualTo(value)
+        Assertions.assertThat(TrialOrder(value).value).isEqualTo(value)
     }
 
     @ParameterizedTest
     @ValueSource(ints = [1, 48, 203802, Int.MAX_VALUE])
     fun `TrialOrder는 객체가 달라도 값이 같으면 동일하다`(value: Int) {
-        assertThat(TrialOrder(value)).isEqualTo(TrialOrder(value))
+        Assertions.assertThat(TrialOrder(value)).isEqualTo(TrialOrder(value))
     }
 
     @ParameterizedTest
     @ValueSource(ints = [0, -1, -278, -2042, Int.MIN_VALUE])
     fun `1부터 순서대로 생성되는 시도 순서는, 양수가 아니면 IllegalArgumentException을 던진다`(value: Int) {
-        assertThatIllegalArgumentException()
+        Assertions.assertThatIllegalArgumentException()
             .isThrownBy { TrialOrder(value) }
             .withMessage("The number must be positive.")
     }
@@ -38,7 +37,7 @@ internal class TrialOrderTest {
     @ValueSource(ints = [1, 48, 203802, Int.MAX_VALUE])
     fun `인덱스가 순서에 맞지 않으면 IllegalArgumentException을 반환한다`(checkingIndex: Int) {
         val orderValue = 10
-        assertThatIllegalArgumentException()
+        Assertions.assertThatIllegalArgumentException()
             .isThrownBy { TrialOrder(orderValue).checkOrder(checkingIndex) }
             .withMessage("The results of racing is not ordered. trialOrder='$orderValue', givenIndex=$checkingIndex")
     }
